chore: rename fns.ts to utils.ts to match frontend convention
