+++ frontend/src/hooks/api/gateways-v2/utils.ts
